sixe

+++ 0x00-hello_world/6-size.c
@@ -10,9 +10,9 @@ int main(void)
 	double doubleType;
 	char charType;
 
-	printf("Size of int: %ld bytes\n", sizeof(integerType));
-	printf("Size of float: %ld bytes\n", sizeof(floatType));
-	printf("Size of double: %ld bytes\n", sizeof(doubleType));
-	printf("Size of char: %ld bytes\n", sizeof(charType));
+	printf("Size of int: %ld bytes\n", (unsigned long)sizeof(integerType));
+	printf("Size of float: %ld bytes\n", (unsigned long)sizeof(floatType));
+	printf("Size of double: %ld bytes\n", (unsigned long)sizeof(doubleType));
+	printf("Size of char: %ld bytes\n", (unsigned long)sizeof(charType));
 	return (0);
 }
